Allow use without numpy (#57)
Conversion of input is evaluated in order of variants declared
in `BytesType`. 

If numpy is not installed, it would error that numpy
was not installed if the input was a Buffer b/c Buffer 
was evaluated after a potential numpy input. 

The patch fixes this behavior by moving the numpy 
variant as the last variant evaluated. Thus wiill only raise 
an error if it's a numpy input/output and numpy is not 
installed.